refactor(searches): modularize conjunction search engine
- Extract find_conjunctions_with_stars to stars.py
- Extract find_conjunctions_between_moving_bodies to pairs.py
- Keep generic find_conjunctions in base.py
- Update package exports and internal references
